Wizard: Rename MountpointSuffix component to MountpointSubpath
ssia
